numTokens, isJson for Tokenizing
Added in a function to return the number of tokens in a JSON string.
This is intended only for performance metrics but you never know.

I also decided to let isJson do the tokenizing (as it already did)
instead of calling tokenize a 2nd time which added time to redo what
was just done.