tokenize() bug with + & -
Found a bug in tokenize() thanks to trying to parse the Livecode
Dictionary and coming across the function “-“. This caused all kinds of
fits as this is a token to Livecode Script.

Added new logic to handle + and - tokens that can determine if the - is
actual data or is part of a number.