Updated my silly little test stack
Lots of changes in this stack. Includes the latest version of fastJson.
I’ve convinced myself that fastJson is faster then both easyJson and
libJson as well as more robust to the point where some of my test files
do not work with either of the older libraries.

As a result, I’ve removed both easyJson and libJson from the test
stack. I’ve also added in radio buttons for JSON validation (self
explanatory I hope).

Also added are 2 text fields to validate library results. First is a
field showing the array created from the JSON file sent through the
library.

The 2nd field shows the resulting JSON created from the array. NOTE:
this WILL NOT be the same as the original JSON. This is due to the way
LC handles arrays.