isNumericalArray() fix
Looking for a fast reliable way to detect numbers as keys. This should
detect only numeric keys and return true only when all keys are a
number. The first time “is a number” false, we return false.

I put “exit isNumericalArray” but I’m not quite sure it’s needed?…